+++ alacritty_terminal/src/event.rs
@@ -1,5 +1,6 @@
 use std::borrow::Cow;
 use std::fmt::{self, Debug, Formatter};
+use std::process::ExitStatus;
 use std::sync::Arc;
 
 use crate::term::ClipboardType;
@@ -53,8 +54,8 @@ pub enum Event {
     /// Shutdown request.
     Exit,
 
-    /// Child process exited with an error code.
-    ChildExit(i32),
+    /// Child process exited.
+    ChildExit(ExitStatus),
 }
 
 impl Debug for Event {
@@ -72,7 +73,7 @@ impl Debug for Event {
             Event::Wakeup => write!(f, "Wakeup"),
             Event::Bell => write!(f, "Bell"),
             Event::Exit => write!(f, "Exit"),
-            Event::ChildExit(code) => write!(f, "ChildExit({code})"),
+            Event::ChildExit(status) => write!(f, "ChildExit({status:?})"),
         }
     }
 }