+++ alacritty_terminal/src/tty/unix.rs
@@ -398,7 +398,7 @@ impl EventedPty for Pty {
                 None
             },
             Ok(None) => None,
-            Ok(exit_status) => Some(ChildEvent::Exited(exit_status.and_then(|s| s.code()))),
+            Ok(exit_status) => Some(ChildEvent::Exited(exit_status)),
         }
     }
 }