+++ alacritty_terminal/CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Pass `-q` to `login` on macOS if `~/.hushlogin` is present
+- **`ChildEvent::Exited` and `Event::ChildExit` now contain `ExitStatus` instead of `i32`**
 
 ## 0.25.0